cleanup, lint and some doc
